Fix permissions issue in Postman tests Dockerfile
When I tested the Dockerfile through WSL, this was not a problem.
But the normal behaviour is that WORKDIR is unaffected by the preceding
USER change.
However, on my system, the created ~/postman-wrapper directory had it's
permissions set to dockeruser.
This change enables 'regular' Linux users (e.g. running in a VM) to also
build and run a container in which permissions are set correctly.